PR#3959 (exit on bad #use in ocaml scripts): a preliminary patch
With this change, the failure of a directive will abort a script if it
is not run in interactive mode. Consider for example the file test.ml:

  #use "!!";;
  print_endline "hello";;

then `ocaml test.ml` will stop after the erroneous #use directive,
while using `#use "test.ml";;` from a running toplevel, or
`ocaml -init test.ml`, will continue the script and print "hello".

The implementation is preliminary: it is minimally invasive, but it
probably makes sense to take this as an opportunity to factorize
error-handling and error-reporting amongst toplevel directives.