ocamldebug: fix debugcomm:Obj.tag on integer values
The debugcomm.ml re-implementation of the Obj module (talking over the
debugging socket) implements Obj.tag without special-casing integers
(that is, testing is_block first), and always assumes its input is
a block. This causes a segfault if it is called on an unboxed integer,
as is the case when printing a lazy value defined as (lazy 1).