Move bytecode debug info from the OCaml heap to static allocations
There is currently a GC bug in the bytecode debug-info handling, due
to the fact that
  void read_main_debug_info(struct debug_info *di)
is passed a internal pointer in the middle of a custom block inside
the OCaml heap. The bug only manifest itself if such custom blocks
are allocated on the minor heap.

This commit fixes this issue by moving debug_info chunks from the
OCaml heap to the C land, stored in a dynamic table. They are
allocated when caml_add_debug_info is called, and removed when
caml_remove_debug_info is called.

It may be the wrong approach, however: maybe we need debug_info to
remain in OCaml land because its lifetime is harder to predict than it
seems (can backtraces point to debug_info referring to a piece of
bytecode that has "removed"?).