menhir parser: remove useless precedence annotation
`%prec below_WITH` must be used at the end of productions that may
have a module_type as their last nonempty nonterminal, and are
themselves module types. Typically,

```
FUNCTOR attributes functor_args MINUSGREATER module_type
      %prec below_WITH
```

needs a precedence because

    functor () -> S with t = bar

could a-priori be parsed as either

    functor () -> (S with t = bar)
    (functor () -> S) with t = bar

Here it was used on the `include S` production, which does not
itself form a module type, so it was not useful.