Revert "parser.mly: use prefixed+currified helpers to facilitate Menhir port"
This reverts commit f0b7535d228c47b4343ab6268c760db9c4912b57.

Those changes to the ocamlyacc parser helped during the yacc->menhir
transition, as they allowed to make the semantic action code
indentical between the yacc and menhir parser, which makes rebasing
against new yacc-parser versions much easier.

Instead of removing the change from the history, we revert it
explicitly so that we can easiy come back later to the non-reverted
version for testing or rebasing.