lambda/switch.ml: pass location information to the make_* combinator parameters
The Switch module is responsible for compiling switches on intervals
of integer values into efficient trees of tests. It is structured as
a functor parametrized on test-generation primitives, to be used by
both the pattern-matching compiler at the Lambda-code level in
lambda/matching.ml and some lower-level Cmm-level operations in
asmcomp/cmmgen.ml

We are working on passing precise location information from the
frontend to the backend. In particular, Switch should track the
location of the input values it is matching over, and assign the
corresponding locations to the tests it generates.

The present commit is a first step in this direction: it adds an
abstract type of locations to the functor argument (to be instantiated
as Location.t for lambda-level switches, Debuginfo.t for cmm-level
swtiches), and updates the users of the functor to this new API. The
compilation logic in Switch itself is unchanged, so for now the
locations being passed are all dummies.

(The `no_loc` argument is hopefully temporary; it should be the case
that, after the Switch code is updated to track locations correctly,
we never need to make up locations during the compilation, so that
`no_loc` can be dropped from the functor interface.)

(This code was actually assembled together by Gabriel Scherer from
Mark's #2294)