avoid physical equality on constant constructors
The camlboot interpreter uses a representation of constant
constructors that does not enforce sharing, and an implementation of
physical equality that checks physical equality of interpreter
values. The two do not combine well.

This is only a partial bandaid for the problem. In the medium term the
camlboot interpreter should be changed to respect physical value
identity during interpretation (this can be defined
clearly/precisely), and think about some guarantees it can give on its
physical equality implementation. I don't know whether
constant-constructors would be covered by those guarantees, though.