Constructor composition in direct style: [benefits_from_dps : bool].
The direct-style version of the tailmap3 function

    let[@tail_mod_cons] rec tailmap3 = fun f li ->
      match li with
      | [] -> []
      | x :: xs -> f 0 x :: f 1 x :: f 2 x :: tailmap3 f xs

looked as follows before this PR:

    (tailmap3/109
       (function f/110 li/111 tail_mod_cons
         (if li/111
           (let (xs/113 =a (field 1 li/111) x/112 =a (field 0 li/111))
             (let (block/115 = (makemutable 0 (apply f/110 0 x/112) 0))
               (seq
                 (let
                   (arg1/116 = (apply f/110 1 x/112)
                    block/117 = (makemutable 0 (apply f/110 2 x/112) 0))
                   (seq
                     (setfield_ptr(heap-init)_computed block/115 1
                       (makeblock 0 arg1/116 block/117))
                     (apply tailmap3_dps/114 block/117 1 f/110 xs/113)))
                 block/115)))
           0a))

It now looks as follows:

    (tailmap3/109
       (function f/110 li/111 trmc
         (if li/111
           (let (xs/113 =a (field 1 li/111) x/112 =a (field 0 li/111))
             (makeblock 0 (apply f/110 0 x/112)
               (makeblock 0 (apply f/110 1 x/112)
                 (let (block/115 = (makemutable 0 (apply f/110 2 x/112) 0))
                   (seq (apply tailmap3_dps/114 block/115 1 f/110 xs/113)
                     block/115)))))

The idea is that the deeply nested constructor (f 2 x :: ...) needs
to be turned into a location for the tailmap3 call to go to the DPS
version, but the code around this constructor itself does not need
to be put in DPS form (as the constructor is already providing
a location). Advertising that it "does not benefit from DPS" let
us generate the much nicer version below, by not trying to create
new locations on each constructor above it.