remove the unused is_native_tail_call_heuristic forward reference
This forward-reference from Lambda to Asmcomp was used to generate
machine-specific tailcall information in -annot output; this only use
was removed in 57d329e07b50e6869875471b085541baadcb8376, so we can now
remove it to simplify the codebase.

The logic was non-trivial and might be useful again in the future.