[WIP] Typedecl_unboxed: new operations on head shapes
We introduce explicitly the "head" of a value; head shapes are finite
approximations of sets of heads. We can compute the head of a value
and check whether a head is included in a given head shape. This will
be useful for toplevel/genprintval.ml which must dynamically decide
the (possibly unboxed) constructor of a dynamic value of known type.

TODO: we should discuss the placement of the types and operations, and
document them;
- maybe move the type of heads in types.mli
- or maybe it's time to introduce dedicated compilation units,
  Head_shape_types (on which Types depends) and Head_shape (on which
  other units depend)?