runtime: allow changing max_domains
caml_max_domains is the maximum number of domain for which a minor
heap area has been reserved. It can be increased, currently by calling
Gc.set, up to a large hardcoded limit Hard_max_domains.

In the future we could dynamically increase caml_max_domains on
demand, at the creation of a new domain.

In the future we could consider doing without Hard_max_domains, but
this requires making various domain-indexed datastructures in the
runtime resizable.

Co-authored-by: Sabine Schmaltz <sabine@tarides.com>