matching debug: do not show default environment of split matrices
When we split a matrix inside a given default environment, the default
environment of the resulting split matrices are highly redundant with
the split matrices themselves: they contain the input default
environment, plus each other.

We hide the default environment of the result matrices to make the
output shorter and less redundant, thus more readable.

Example before:

  SPLIT
    _
    0
    1
    3
    _
  INTO:
    First matrix:
      PM:
        _
      Default environment:
        Matrix for 2:
          <0>
          <1>
          <3>
        Matrix for 1:
          <_>
    Default matrix 2:
      PM:
        0
        1
        3
      Default environment:
        Matrix for 1:
          <_>
    Default matrix 1:
      PM:
        _

Notice that the "default matrix 2" and "default matrix 1" are repeated
in the default environment of the "First matrix".

Example after:

  SPLIT
    _
    0
    1
    3
    _
  INTO:
    First matrix:
      PM:
        _
    Default matrix 2:
      PM:
        0
        1
        3
    Default matrix 1:
      PM:
        _