matching debug: show record patterns whose fields are wildcards
Before this PR, a record pattern that accepts all values for all its
fields will be pretty-printed (in exhaustivity counter-examples and in
the debug output of the pattern-matching compiler0 as a wildcard
pattern `_`. This is correct for users, but it gives inaccurate
information in the debug output, as the two patterns are distinct.

Instead I propose to show "{ _ }" in this case, to mean a record whose
field values do not matter. This happens to not be valid OCaml syntax,
but it is perfectly understandable (it is the 0-ary version of the
{ <fields>; _ } syntax) and thus unlikely to confuse users.