matching: specialized handling of the global 'fail exit'
This approach should preserve the current compilation behavior. It is
more efficient for [Partial] matches, as we avoid tracking any context
information on the final exit.

We hope that this refined interface will also let us implement a fine-grained
fix for totality failures due to mutable value. The idea is that when
we detect that some current values may flow to the final exit, we
could sometimes weaken [Jumps.empty Total] into [Jumps.empty Partial]
if we are under a mutable constructor.