remove the previous check_partial stuff
The previous [check_partial] stuff was a coarse-grained heuristic to
degrade matching totality information in certain cases. It was not
quite correct (it failed to prevent The Bug but also instances of it
that it should have caught), and is not sufficient conceptually for
OCaml Multicore. In a Multicore world, other domains may race to
update values concurrently, so we cannot assume that only a fixed set
of matching constructions may result in side-effects.