[refactor] minor renaming in lambda/matching.ml
The name `transl_option` indicates that it matches on option values,
but `combine_constructor` would use it in other cases as well (for
example, on lists).

To reason about the `combine_constructor`, we need to know how
to it is implemented: would it be valid to still use the function in
matches with several non-constant constructors that all have the same
action?

The present refactoring clarifies, in the context of
`combine_constructor`, precisely which check is being implemented, to
make it easier to reason on the code and extend it later.